requirements.txt
streamlit
pytesseract
Pillow
pdfplumber